+++ flake.nix
@@ -3,16 +3,18 @@
 
   inputs = {
     nixpkgs.url = "github:nixos/nixpkgs/nixos-22.11";
+    import-cargo.url = github:edolstra/import-cargo;
     flake-utils.url = "github:numtide/flake-utils";
   };
 
-  outputs = { self, nixpkgs, flake-utils }:
+  outputs = { self, nixpkgs, import-cargo, flake-utils }:
     {
       overlays = rec {
         mineflake = final: prev: {
           mineflake = import ./pkgs {
             pkgs = prev;
             lib = prev.lib;
+            inherit (import-cargo.builders) importCargo;
           };
         };
 
@@ -20,6 +22,7 @@
           mineflake = import ./pkgs ({
             pkgs = prev;
             lib = prev.lib;
+            inherit (import-cargo.builders) importCargo;
           } // attrs);
         };